# backend/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "BandMate7 API",
        "contact": {}
    },
    "paths": {
...
        "/performances": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performances"
                ],
                "summary": "Get all performances",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Desc",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order By",
                        "name": "orderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Performance"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performances"
                ],
                "summary": "Create a new performance",
                "parameters": [
                    {
                        "maxLength": 255,
                        "minLength": 1,
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Bpm",
                        "name": "bpm",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Album/Performance cover",
                        "name": "cover",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Upload file",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Set user role",
                        "name": "userRoleId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Performance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/performances/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performances"
                ],
                "summary": "Get performance by id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Performance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Performance"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performances"
                ],
                "summary": "Delete a performances",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Performance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/performances/{id}/cover": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "performances"
                ],
                "summary": "Set a cover for a performance",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Performance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Album/Performance cover",
                        "name": "cover",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
        "/performances/{id}/resources": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performances"
                ],
                "summary": "Get all resources from a performance",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Performance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Resource"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performances"
                ],
                "summary": "Create a new resource",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Performance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Role",
                        "name": "role",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Bild",
                        "name": "resource",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Resource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Performance": {
            "type": "object",
            "properties": {
                "artist": {
                    "$ref": "#/definitions/model.Artist"
                },
                "bpm": {
                    "type": "integer"
                },
                "cover": {
                    "$ref": "#/definitions/Resource"
                },
                "createdAt": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Genre"
                    }
                },
                "id": {
                    "type": "string"
                },
                "meter": {
                    "$ref": "#/definitions/model.Meter"
                },
                "released": {
                    "type": "integer"
                },
                "resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Resource"
                    }
                },
                "titel": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "Resource": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_role": {
                    "$ref": "#/definitions/UserRole"
                }
            }
        },
        "UserRole": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "titel": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.Artist": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "titel": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Genre": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "titel": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Meter": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "titel": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}